+++ js/shooter.js
@@ -28,7 +28,7 @@ var App = function() {
 	var fireRateTemp   = 4;							// Temporary fire rate (e.g. overrides default fire rate if cheat-mode is enabled).
 	var fireDamage     = 200;						// How much damage is caused by one shot of the player's ship.
 	
-	var enemyHealth    = [ 400, 800, 1200, 1600 ];	// Enemy's health.
+	var enemyHealth    = [ 400, 600, 800, 1500, 5000 ];	// Enemy's health.
 	var enemyDelay     = 2000;						// How long to wait from spawning one enemy to spawning the next one.
 	var nextEnemy;									// The process that will spawn the next enemy.
 
@@ -46,6 +46,8 @@ var App = function() {
 	
 	var levelCounter;								// Object to display the level.
 	var level          = 1;							// Initial level.
+	var loopUid        = null;						// Uid of the background music. Used to start/stop it.
+	var musicPlaying   = false;						// Status of background music
 	
 	var images = {
 		logo		: '../img/logo.png',
@@ -74,15 +76,42 @@ var App = function() {
 		},
 		
 		// Bullets
-		bullet      : '../img/bullets/bullet2.png',
-		enemyBullet : '../img/bullets/bullet3.png',
+		shipBullet   : '../img/bullets/bullet_ship.png',
+		enemyBullets : {
+			0 : {
+					file: '../img/bullets/bullet0.png',
+					delay: 600,
+					damage: 1
+				},
+			1 : {
+					file: '../img/bullets/bullet1.png',
+					delay: 500,
+					damage: 3
+				},
+			2 : {
+					file: '../img/bullets/bullet2.png',
+					delay: 400,
+					damage: 5
+				},
+			3 : {
+					file: '../img/bullets/bullet3.png',
+					delay: 300,
+					damage: 10
+				},
+			4 : {
+					file: '../img/bullets/bullet4.png',
+					delay: 200,
+					damage: 20
+				}
+		},
 		
 		// Enemies
 		enemies     : {
 			0       : '../img/enemies/enemy0.png',
 			1       : '../img/enemies/enemy1.png',
 			2       : '../img/enemies/enemy2.png',
-			3       : '../img/enemies/enemy3.png'
+			3       : '../img/enemies/enemy3.png',
+			4       : '../img/enemies/enemy4.png'
 		},
 		
 		// Asteroids
@@ -108,13 +137,15 @@ var App = function() {
 	var sounds = {
 		shoot       : '../sounds/shoot.mp3',
 		hit         : '../sounds/hit.mp3',
-		explode     : '../sounds/explode.mp3'
+		explode     : '../sounds/explode.mp3',
+		loop        : '../sounds/loop.mp3'
 	}
 	
-	var toggleTitle       = document.getElementById('toggleTitle');
-	var toggleRendererBtn = $('#toggleRendererBtn');
+	var toggleRendererTitle    = document.getElementById('toggleRendererTitle');
+	var toggleRendererBtn      = $('#toggleRendererBtn');
+	var toggleMusicTitle       = document.getElementById('toggleMusicTitle');
+	var toggleMusicBtn         = $('#toggleMusicBtn');
 
-	
 	/**
 	 * Load images and sounds. Also set screen size.
 	 */
@@ -122,11 +153,14 @@ var App = function() {
 		// Images
 		wade.loadImage(images.logo);
 		wade.loadImage(images.ship);
-		wade.loadImage(images.bullet);
-		wade.loadImage(images.enemyBullet);
+		
+		// Bullets
+		wade.loadImage(images.shipBullet);
+		for (var i = 0; i < Object.keys(images.enemyBullets).length; i++) {
+			wade.loadImage(images.enemyBullets[i].file);
+		}
 		
 		// Animation
-		//wade.loadImage(images.boom);
 		for (var i = 0; i < Object.keys(images.boom).length; i++) {
 			wade.loadImage(images.boom[i].file);
 		}
@@ -150,6 +184,7 @@ var App = function() {
 		wade.loadAudio(sounds.shoot);
 		wade.loadAudio(sounds.hit);
 		wade.loadAudio(sounds.explode);
+		wade.loadAudio(sounds.loop);
 	};
 
 	
@@ -166,21 +201,38 @@ var App = function() {
 			wade.setLayerRenderMode(defaultLayerId, '2d');
 			toggleRendererBtn.removeClass('fa-toggle-on');
 			toggleRendererBtn.addClass('fa-toggle-off');
-			toggleTitle.title = 'Enable WebGL';
+			toggleRendererTitle.title = 'Enable WebGL';
 		} else {
 			wade.setLayerRenderMode(defaultLayerId, 'webgl');
 			toggleRendererBtn.removeClass('fa-toggle-off');
 			toggleRendererBtn.addClass('fa-toggle-on');
-			toggleTitle.title = 'Disable WebGL';
+			toggleRendererTitle.title = 'Disable WebGL';
+		}
+
+		
+		// Check whether to play background music.
+		var musicPlayingData = wade.retrieveLocalObject('music');
+		musicPlaying = (musicPlayingData && musicPlayingData.music) || musicPlaying;
+		musicPlaying = (getCookie('music') == 'true') || musicPlaying;
+		if (musicPlaying) {
+			toggleMusicBtn.removeClass('music-off');
+			toggleMusicBtn.addClass('music-on');
+			toggleMusicTitle.title = 'Disable music';
 		}
 
 		
 		// Set screen size to current size of viewport.
+		// TODO: testing
+		wade.setResolutionFactor(1);
+		
 		wade.setMinScreenSize($(window).width(), $(window).height());
 		wade.setMaxScreenSize($(window).width(), $(window).height());
 		//console.log('Layer render mode: ' + wade.getLayerRenderMode(defaultLayerId) + '\nforce2d: ' + force2d + '\nScreen size set to: ' + $(window).width() + 'x' + $(window).height());
 
+		
+		// Get default renderer.
 		var defaultRenderer = wade.getLayerRenderMode(defaultLayerId);
+		
 		
 		// Load highscore.
 		var shooterData = wade.retrieveLocalObject('shooterData');
@@ -206,19 +258,19 @@ var App = function() {
 		}
 		
 		
-		/**
-		 * Main screen text.
-		 */
+		// Main screen text.
 		var menuTexts = {
 				insertCoin  : '- INSERT COIN -',
 				highscoreIs : 'HIGHSCORE IS %i POINTS',
 				youScored   : 'YOU %s %i POINTS'
 		}
 		
+		
 		// Ugly workaround for cut-off texts when using WebGL.
 		if (defaultRenderer == 'webgl') {
 			menuTexts = padStrings(menuTexts);
 		}
+		
 		
 		// Prepare text sprites of main screen.
 		var clickText = new TextSprite(menuTexts.insertCoin, '36pt Highspeed', 'white', 'center');
@@ -265,6 +317,7 @@ var App = function() {
 			clickToStart.addSprite(new TextSprite(menuTexts.highscoreIs, '24pt Highspeed', 'yellow', 'center'), { y: 160 });
 		}
 
+		
 		// Store highscore only if player didn't cheat. 
 		if(!cheat) {
 			// Update highscore cookie.
@@ -275,8 +328,10 @@ var App = function() {
 			wade.storeLocalObject('shooterData', shooterData);
 		}
 		
+		
 		// Add logo
 		clickToStart.addSprite(new Sprite(images.logo), { y: -200 });
+		
 		
 		// Add pause text into hidden scene object.
 		pauseSpriteObj = new SceneObject();
@@ -285,14 +340,34 @@ var App = function() {
 		pauseSpriteObj.addSprite(new Sprite(images.logo), { y: -200 });
 		pauseSpriteObj.addSprite(pauseSpriteText, { y: 160 });
 		
+		
 		// Show close button and default cursor while not playing.
 		gameBtnObj.style.display = 'block';
 		game.style.cursor = 'default';
 		
+		
 		// Show main menu.
 		wade.addSceneObject(clickToStart);
 		
+		
+		// Initialize asteroids on the main screen.
 		nextAsteroid = setTimeout(wade.app.spawnAsteroid, asteroidDelay);
+		
+		
+		// Decide whether to play music or not.
+		if (musicPlaying) {
+			loopUid = wade.playAudio(sounds.loop, true);
+			
+			// On error
+			if (loopUid < 0) {
+				toggleMusicBtn.removeClass('music-on');
+				toggleMusicBtn.addClass('music-off');
+				toggleMusicTitle.title = 'Enable music';
+				musicPlaying = false;
+			}
+		}		
+		//console.log('Uid: ' + loopUid + ', music: ' + musicPlaying);
+
 		
 		/**
 		 * Start game on left mouse click.
@@ -344,7 +419,7 @@ var App = function() {
 				lastFireTime = time;
 				var shipPosition = ship.getPosition();
 				var shipSize = ship.getSprite().getSize();
-				var sprite = new Sprite(images.bullet);
+				var sprite = new Sprite(images.shipBullet);
 				var bullet = new SceneObject(sprite, 0, shipPosition.x, shipPosition.y - shipSize.y / 2);
 				wade.addSceneObject(bullet);
 				wade.playAudio(sounds.shoot, false);
@@ -364,7 +439,7 @@ var App = function() {
 					score -= 10;
 				}
 				
-				if (score < 0) {
+				if (score < 0 || score == 'NaN') {
 					score = 0;
 				}
 			}
@@ -399,7 +474,7 @@ var App = function() {
 									wade.removeSceneObject(colliders[j]);
 
 									// Increase score if enemy/asteroid shot down by a 10th of its initial health.
-									score += Math.floor(colliders[j].initialHealth / 10);
+									score += Math.floor(colliders[j].initialHealth / 10) * level;
 								}
 
 								// Delete bullet.
@@ -414,9 +489,11 @@ var App = function() {
 			}
 
 			// Increase level and health every 1000 points.
-			if (Math.floor(score / 1000) > level) {
+			if (Math.floor(score / 1000) > level && level < 5) {
 				level += 1;
-				playerHealth += 10;
+				if (!cheat) {
+					playerHealth = 100;
+				}
 
 				// Make enemies spawn faster.
 				if (enemyDelay > 200) {
@@ -453,30 +530,38 @@ var App = function() {
 			// Get a list of overlapping objects on the current layer.
 			var overlapping = ship.getOverlappingObjects(false, 'axis-aligned');
 			var hit = false;
+			var enemyDamage = 1;
 			
 			// If the list is not empty ...
 			if (overlapping.length > 0) {
 				
 				for (var i = 0; i < overlapping.length; i++) {
 
-					// ... check if the overlapping object is either an enemy or an enemy's bullet.
-					if (overlapping[i].isEnemy || overlapping[i].isEnemyBullet) {
-						
-						// Comparing per-pixel is quite slow, but the only easy way to check for collisions while discarding transparent pixels.
-						if (ship.overlapsObject(overlapping[i], 'pixel')) {
-							// Decrease health of overlapping object by a 10th of the default fire damage.
-							if (overlapping[i].health > 0) {
-								overlapping[i].health -= Math.floor(fireDamage / 10);
-							}
+					if (typeof(overlapping[i]) != 'undefined') {
+						// ... check if the overlapping object is either an enemy or an enemy's bullet.
+						if (overlapping[i].isEnemy || overlapping[i].isEnemyBullet) {
 							
-							// Remove enemy's bullet if it hit the player's ship.
-							if (overlapping[i].isEnemyBullet || overlapping[i].health <= 0) {
-								wade.removeSceneObject(overlapping[i]);
-								wade.removeObjectFromArrayByIndex(i, overlapping);
+							// Comparing per-pixel is quite slow, but the only easy way to check for collisions while discarding transparent pixels.
+							if (typeof(overlapping[i]) != 'undefined' && ship.overlapsObject(overlapping[i], 'pixel')) {
+								// Decrease health of overlapping object by a 10th of the default fire damage.
+								if (overlapping[i].health > 0) {
+									overlapping[i].health -= Math.floor(fireDamage / 10);
+								}
+								
+								// Remove enemy's bullet and/or ship if it hit the player's ship.
+								if (typeof(overlapping[i]) != 'undefined' && (overlapping[i].isEnemyBullet || overlapping[i].health <= 0)) {
+									wade.removeSceneObject(overlapping[i]);
+									wade.removeObjectFromArrayByIndex(i, overlapping);
+								}
+								
+								if (typeof(overlapping[i]) != 'undefined') {
+									enemyDamage = overlapping[i].damage || enemyDamage;
+								}
+
+								//console.log(enemyDamage);
+								hit = true;
+								break;
 							}
-							
-							hit = true;
-							break;
 						}
 					}
 				}
@@ -491,7 +576,11 @@ var App = function() {
 
 				// Decrease health.
 				if (playerHealth > 0) {
-					playerHealth--;
+					playerHealth -= enemyDamage;
+				}
+				
+				if (playerHealth < 0 || playerHealth == 'NaN') {
+					playerHealth = 0;
 				}
 				healthCounter.getSprite().setText(playerHealth);
 				
@@ -708,6 +797,7 @@ var App = function() {
 		asteroid.isAsteroid = true;
 		asteroid.health = asteroidHealth[asteroidId];
 		asteroid.initialHealth = asteroidHealth[asteroidId];
+		asteroid.damage = Math.floor(asteroidHealth[asteroidId] / 100);
 		
 		
 		/**
@@ -729,8 +819,18 @@ var App = function() {
 		// Create an empty sprite.
 		var sprite;
 
+		var maxEnemy = level - 1;
+		var enemyCount = Object.keys(images.enemies).length - 1;
+		
+		//console.log('enemies: ' + enemyCount);
+		//console.log('level: ' + level);
+
+		if (maxEnemy > enemyCount) {
+			maxEnemy = enemyCount;
+		}
+		
 		// Select random image of enemy as sprite.
-		var enemyId = getRandomInt(0, 3); 
+		var enemyId = getRandomInt(0, maxEnemy);
 		sprite = new Sprite(images.enemies[enemyId]);
 
 		// Calculate start and end coordinates.
@@ -747,6 +847,7 @@ var App = function() {
 		enemy.isAsteroid = false;
 		enemy.health = enemyHealth[enemyId];
 		enemy.initialHealth = enemyHealth[enemyId];
+		enemy.damage = Math.floor(enemyHealth[enemyId] / 100);
 
 		
 		/**
@@ -792,9 +893,10 @@ var App = function() {
 			var endY = startY + dy * 3000;
 
 			// Create bullet.
-			var sprite = new Sprite(images.enemyBullet);
+			var sprite = new Sprite(images.enemyBullets[enemyId].file);
 			var bullet = new SceneObject(sprite, 0, startX, startY);
 			bullet.isEnemyBullet = true;
+			bullet.damage = images.enemyBullets[enemyId].damage;
 			wade.addSceneObject(bullet);
 			bullet.moveTo(endX, endY, 200);
 
@@ -804,7 +906,7 @@ var App = function() {
 			};
 
 			// Schedule next bullet.
-			this.schedule(600, 'fire'); // 1000
+			this.schedule(images.enemyBullets[enemyId].delay, 'fire'); // 1000
 		};
 		enemy.schedule(300, 'fire'); // 500
 		
@@ -833,6 +935,31 @@ var App = function() {
 		wade.clearScene();
 		wade.app.init();
 	};
+
+	
+	/**
+	 * Toogle background music
+	 */
+	this.toggleMusic = function() {
+		if (musicPlaying) {
+			if (loopUid > -1) {
+				wade.stopAudio(loopUid);
+			}
+			toggleMusicBtn.removeClass('music-on');
+			toggleMusicBtn.addClass('music-off');
+			toggleMusicTitle.title = 'Enable music';
+		} else {
+			loopUid = wade.playAudio(sounds.loop, true);
+			toggleMusicBtn.removeClass('music-off');
+			toggleMusicBtn.addClass('music-on');
+			toggleMusicTitle.title = 'Disable music';
+		}
+		
+		musicPlaying = !musicPlaying;
+		var shooterData = { music: musicPlaying };
+		wade.storeLocalObject('music', shooterData);
+		setCookie('music', musicPlaying, 365);
+	}
 };
 
 
@@ -853,7 +980,7 @@ function getRandomInt(min, max) {
  * @returns entity of input type
  */
 function padStrings(obj) {
-	var padding = '       ';
+	var padding = '       ';
 	if (typeof(obj) == 'object') {
 		var tmp = {};
 		for (var i in obj) {